requirements.txt
django==4.2.9
djangorestframework==3.14.0
python-dotenv==1.0.0
openai==1.6.1
twilio==8.10.0
requests==2.31.0
pydub==0.25.1
celery==5.3.4
redis==5.0.1
whitenoise==6.9.0
django-cors-headers==4.7.0